requirements.txt
requests>=2.31,<3
pymongo>=4.6,<5
schedule>=1.2,<2
python-dotenv>=1.0,<2
